P0452R1 Unifying <numeric> Parallel Algorithms
This paper incorporates the pending changes from the
next paper to commit, P0467R2, to the modified algorithms,
and this commit follows suit.  We choose to apply papers
in this order to minimize overall churn.

This branch is based on the p0523r1 branch, and is third
in a sequence of six commits for parallel algorithms.