P0467R2 Iterator Concerns for Parallel Algorithms
Minor editorial tweaks splitting uo a few paragraphs for clarity,
when dealing with both serial and parallel algorithms in wordy
clauses.

Added extra section-header comments consistently through the
<algorithm> and <numeric> headers to make navigation simpler
while validating the changes.

This branch is based on the p0452r1 branch, and is fourth in a
sequence of six commits for parallel algorithms.